# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Monitor PTAP API
  version: 1.0.0
  description: >
    Backend NestJS del monorepo monitor-ptap: gateway OPC UA de 13 PTAP.
    Prefijo global /api (excepto /metrics, fuera del prefijo por convención de Prometheus).
    Puerto 4000 (variable PORT).

    Autorización por PERMISOS granulares (@ptap/shared → ROLE_PERMISSIONS/hasPermission), no por
    roles lineales: sin token → 401; sin el permiso requerido → 403. Roles:
    civil (vista básica) | operador (datos + válvulas) | jefe (datos, sin válvulas) | admin (total).

    Tiempo real por Socket.IO: el cliente emite 'opc:subscribe' { plantId }; el servidor emite
    'opc:snapshot' (por planta, solo en cambios) y 'opc:liveness' (broadcast).
    NOTA: el gateway Socket.IO aún NO autentica el handshake (ver docs/SECURITY_FINDING_P0.md §6).
servers:
  - url: http://localhost:4000
    description: Backend local (dev)
...
paths:
...
  /api/auth/login:
    post:
      summary: Iniciar sesión
      operationId: login
      tags: [auth]
      security: []
      description: Rate-limit propio, más estricto que el global (LOGIN_RATE_LIMIT_*).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '201': { $ref: '#/components/responses/AuthResult' }
        '401': { description: Credenciales inválidas }
        '429': { description: Demasiados intentos (rate-limit) }
  /api/auth/register:
    post:
      summary: Auto-registro (crea SIEMPRE rol civil)
      operationId: register
      tags: [auth]
      security: []
      description: >
        La cuenta nace SIEMPRE con rol 'civil' (solo consulta): el rol lo fija el servidor.
        El schema es .strict(), así que enviar 'role' en el body devuelve 400 — nadie puede
        auto-asignarse un rol. Elevarlo es potestad del Administrador (PATCH /api/users/{id}/role).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, plant, password]
              additionalProperties: false
              properties:
                name: { type: string, minLength: 2, maxLength: 120 }
                email: { type: string, format: email }
                phone: { type: string, minLength: 5, maxLength: 32 }
                plant: { type: string, description: Slug canónico de la planta, example: montebello }
                password: { type: string, minLength: 8 }
      responses:
        '201': { $ref: '#/components/responses/AuthResult' }
        '400': { description: Body inválido (incluye el caso de enviar 'role') }
        '409': { description: Ese correo ya está registrado }
        '429': { description: Demasiados intentos (rate-limit) }

  # ── USERS (solo Admin) ──────────────────────────────────────────────────────
  /api/users:
    get:
      summary: Listar usuarios
      operationId: listUsers
      tags: [users]
      description: Requiere el permiso `manage_users` (solo admin). Nunca expone password_hash.
      responses:
        '200':
          description: Lista de usuarios
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/UserSummary' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /api/users/{id}/role:
    patch:
      summary: Asignar rol a un usuario
      operationId: updateUserRole
      tags: [users]
      description: >
        Requiere el permiso `assign_roles` (solo admin). Audita el cambio (quién, a quién, de→a).
        Guard rail: un admin NO puede cambiar su propio rol (400).
        El rol viaja en el JWT (8 h): aplica cuando el usuario vuelve a iniciar sesión.
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [role]
              additionalProperties: false
              properties:
                role: { $ref: '#/components/schemas/Role' }
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '400': { description: Rol inválido, o intento de cambiar el propio rol }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Usuario no encontrado }
  /api/users/{id}/active:
    patch:
      summary: Activar/desactivar una cuenta
      operationId: updateUserActive
      tags: [users]
      description: >
        Requiere el permiso `manage_users` (solo admin). Un usuario inactivo no puede iniciar
        sesión. Guard rail: un admin NO puede desactivarse a sí mismo (400).
      parameters: [{ $ref: '#/components/parameters/UserId' }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [isActive]
              additionalProperties: false
              properties:
                isActive: { type: boolean }
      responses:
        '200':
          description: Usuario actualizado
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UserSummary' }
        '400': { description: Intento de desactivarse a sí mismo }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Usuario no encontrado }

  # ── PLANTS (telemetría) ─────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    UserId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Unauthorized:
      description: Falta el JWT o no es válido
    Forbidden:
      description: El rol no tiene el permiso requerido
    AuthResult:
      description: Token JWT + usuario
      content:
        application/json:
          schema:
            type: object
            properties:
              token: { type: string, description: JWT (expira según JWT_EXPIRES_IN, default 8h) }
              user: { $ref: '#/components/schemas/AuthUser' }
  schemas:
    Role:
      type: string
      enum: [civil, operador, jefe, admin]
    AuthUser:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/Role' }
        plant: { type: string, example: montebello }
    UserSummary:
      type: object
      description: Usuario expuesto por la API de administración. NUNCA incluye password_hash.
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        phone: { type: string, nullable: true }
        role: { $ref: '#/components/schemas/Role' }
        plant: { type: string }
        isActive: { type: boolean }
        lastLoginAt: { type: string, format: date-time, nullable: true }
        createdAt: { type: string, format: date-time, nullable: true }
...
security:
  - bearerAuth: []